Switch to centipawn evaluation and use timer to track search time

+++ src/evaluate.rs
@@ -3,5 +3,5 @@ use super::board::Board;
 pub mod simple_evaluator;
 
 pub trait Evaluator: Clone {
-    fn evaluate(&self, board: &Board) -> i32;
+    fn evaluate(&self, board: &Board) -> i64;
 }
